Add getNumberOfAds()
getNumberOfAds() is implemented in all layers.

Instead of requesting the entire Ad table and manipulate the data on the
client side, getNumberOfAds() only requests the count of all ads for a given job id. This is used for displaying the correct number of ad tabs, as well as hiding the ad component if there are no ads.