feat 10: getJobById()
findById() for jobs returns the entire entity. Filtering the entire
table in backend to a DTO is an antipattern (or so I heard).

Using JPQL query, only the relevant columns are returned. The Optional
return is used so that existing exception handling can be used. The existing
JobDtoView is slimmed to only title and description. The old findById()
is kept for testing purposes.