feat: getEmailUnique
When a user tries to register a new Email, that email cannot already exist in the db.

The user sends a request to account/checkEmailUnique/{email}. If the email is not already registered, the response includes a true flag. If the email is already registered, a EmailNotFreeException() will get thrown.
Method added to repo, service, and controller layer. Tests for each layer.