feat: verification email with code
Added Verification class, with repository and service layer. When a new
email is registered, a verification object is stored in the table
"verification". Verification knows about the account, but the account
doesn't know about the verification object.

Added mock tests to the service layer.

When a new email is registered, an email with a verification code is
sent to that email.